feat: ApiRequestError for unsuccessful requests
If there is an error with a request, Shipstation can respond with a
simple error string rather than JSON. This causes a JSON::ParserError
and so, while the Shipstation error message is accessable by rescuing
JSON::ParserError, there is no way to access the response instance.

If the error is 429 (Too Many Requests), Shipstation sets the number of
seconds to wait until the rate limit is reset. By raising an
ApiRequestError which contains the request headers, it makes it possible
sleep an application until Shipstation is ready to accept requests again